+++ classes/local/logbook/data_access/logbook_vault_interface.php
@@ -39,7 +39,7 @@ interface logbook_vault_interface {
      * @param logbook   $logbook    The logbook_interface of for all entries.
      * @return logentries[]     Array of logentry_interfaces.
      */
-    public function get_logbook(int $courseid, int $studentid, $logbook = null);
+    public static function get_logbook(int $courseid, int $studentid, $logbook = null);
 
     /**
      * Get a specific logbook entry.
@@ -51,7 +51,16 @@ interface logbook_vault_interface {
      * @param logbook   $logbook    The logbook_interface of for all entries.
      * @return logentry         A logentry_insterface.
      */
-    public function get_logentry(int $studentid, int $courseid, int $logentryid = 0, int $exerciseid = 0, $logbook);
+    public static function get_logentry(int $studentid, int $courseid, int $logentryid = 0, int $exerciseid = 0, $logbook);
+
+    /**
+     * Update a student's logbook entry
+     *
+     * @param int       $courseid   The course id associated with the logbook.
+     * @param int       $studentid  The student id associated with the logbook.
+     * @return array    $totalflighttime, $totalsessiontime, $totalsolotime
+     */
+    public static function get_logbook_summary(int $courseid, int $studentid);
 
     /**
      * Create a student's logbook entry
@@ -61,7 +70,7 @@ interface logbook_vault_interface {
      * @param logentry  $logentry A logbook entry of the student.
      * @return bool     $result of the database add operation.
      */
-    public function insert_logentry(int $courseid, int $studentid, logentry $logentry);
+    public static function insert_logentry(int $courseid, int $studentid, logentry $logentry);
 
     /**
      * Update a student's logbook entry
@@ -71,7 +80,7 @@ interface logbook_vault_interface {
      * @param logentry  $logentry A logbook entry of the student.
      * @return bool     $result of the database update operation.
      */
-    public function update_logentry(int $courseid, int $studentid, logentry $logentry);
+    public static function update_logentry(int $courseid, int $studentid, logentry $logentry);
 
     /**
      * Delete a logbook entry by id
@@ -79,5 +88,5 @@ interface logbook_vault_interface {
      * @param int   $logentryid   The logbook entry id to be deleted.
      * @return bool result of the database update operation.
      */
-    public function delete_logentry($logentryid);
+    public static function delete_logentry($logentryid);
 }