+++ classes/local/session/data_access/analytics_vault_interface.php
@@ -36,7 +36,7 @@ interface analytics_vault_interface {
      * @param int   $studentid  The student id in reference
      * @return int  $days       The number of days since last session
      */
-    public function get_session_recency(int $courseid, int $studentid);
+    public static function get_session_recency(int $courseid, int $studentid);
 
     /**
      * Get the number of Availability slots marked by the student.
@@ -44,7 +44,7 @@ interface analytics_vault_interface {
      * @param int   $studentid  The student id in reference
      * @return int  $slotcount  The number of availability slots marked by the student.
      */
-    public function get_slot_count(int $studentid);
+    public static function get_slot_count(int $studentid);
 
     /**
      * Get course activity for a student from the logs.
@@ -52,7 +52,7 @@ interface analytics_vault_interface {
      * @param int   $studentid      The student id in reference
      * @return int  $activitycount  The number of activity events in the log.
      */
-    public function get_activity_count(int $studentid);
+    public static function get_activity_count(int $studentid);
 
     /**
      * Get course activity for a student from the logs.
@@ -60,5 +60,5 @@ interface analytics_vault_interface {
      * @param int   $studentid      The student id in reference
      * @return int  $completions    The number of lesson completions.
      */
-    public function get_lesson_completions(int $studentid);
+    public static function get_lesson_completions(int $studentid);
 }